+++ src/Propulsion.Cosmos/CosmosSink.fs
@@ -26,7 +26,7 @@ module Internal =
             | Duplicate of updatedPos : int64
             | PartialDuplicate of overage : StreamSpan<byte[]>
             | PrefixMissing of batch : StreamSpan<byte[]> * writePos : int64
-        let logTo (log : ILogger) (res : StreamName * Choice<EventStats * Result, EventStats * exn>) =
+        let logTo (log : ILogger) (res : StreamName * Choice<EventMetrics * Result, EventMetrics * exn>) =
             match res with
             | stream, (Choice1Of2 (_, Ok pos)) ->
                 log.Information("Wrote     {stream} up to {pos}", stream, pos)
@@ -80,7 +80,7 @@ module Internal =
             | ResultKind.TooLarge | ResultKind.Malformed -> true
 
     type CosmosStats(log : ILogger, statsInterval, stateInterval) =
-        inherit Scheduling.StreamSchedulerStats<EventStats * Writer.Result, EventStats * exn>(log, statsInterval, stateInterval)
+        inherit Scheduling.Stats<EventMetrics * Writer.Result, EventMetrics * exn>(log, statsInterval, stateInterval)
         let okStreams, resultOk, resultDup, resultPartialDup, resultPrefix, resultExnOther = HashSet(), ref 0, ref 0, ref 0, ref 0, ref 0
         let badCats, failStreams, rateLimited, timedOut, tooLarge, malformed = CatStats(), HashSet(), ref 0, ref 0, ref 0, ref 0
         let rlStreams, toStreams, tlStreams, mfStreams, oStreams = HashSet(), HashSet(), HashSet(), HashSet(), HashSet()
@@ -131,7 +131,7 @@ module Internal =
     type CosmosSchedulingEngine =
 
         static member Create(log : ILogger, cosmosContexts : _ [], itemDispatcher, stats : CosmosStats, dumpStreams, ?maxBatches)
-            : Scheduling.StreamSchedulingEngine<_, _> =
+            : Scheduling.StreamSchedulingEngine<_, _, _> =
             let writerResultLog = log.ForContext<Writer.Result>()
             let mutable robin = 0
             let attemptWrite (item : Scheduling.DispatchItem<_>) = async {
@@ -153,8 +153,8 @@ module Internal =
                         streams.SetMalformed(stream,malformed)
                 let _stream, ss = applyResultToStreamState res
                 Writer.logTo writerResultLog (stream,res)
-                ss.write
-            let dispatcher = Scheduling.MultiDispatcher<_, _>(itemDispatcher, attemptWrite, interpretWriteResultProgress, stats, dumpStreams)
+                ss.write, res
+            let dispatcher = Scheduling.MultiDispatcher<_, _, _>(itemDispatcher, attemptWrite, interpretWriteResultProgress, stats, dumpStreams)
             Scheduling.StreamSchedulingEngine(dispatcher, enableSlipstreaming=true, ?maxBatches = maxBatches)
 
 type CosmosSink =