+++ src/Propulsion.EventStore/EventStoreSink.fs
@@ -25,7 +25,7 @@ module Internal =
             | PartialDuplicate of overage : StreamSpan<byte[]>
             | PrefixMissing of batch : StreamSpan<byte[]> * writePos : int64
 
-        let logTo (log : ILogger) (res : FsCodec.StreamName * Choice<EventStats * Result, EventStats * exn>) =
+        let logTo (log : ILogger) (res : FsCodec.StreamName * Choice<EventMetrics * Result, EventMetrics * exn>) =
             match res with
             | stream, (Choice1Of2 (_, Ok pos)) ->
                 log.Information("Wrote     {stream} up to {pos}", stream, pos)
@@ -65,7 +65,7 @@ module Internal =
             | _ -> ResultKind.Other
 
     type EventStoreStats(log : ILogger, statsInterval, stateInterval) =
-        inherit Scheduling.StreamSchedulerStats<EventStats * Writer.Result, EventStats * exn>(log, statsInterval, stateInterval)
+        inherit Scheduling.Stats<EventMetrics * Writer.Result, EventMetrics * exn>(log, statsInterval, stateInterval)
         let okStreams, resultOk, resultDup, resultPartialDup, resultPrefix, resultExnOther = HashSet(), ref 0, ref 0, ref 0, ref 0, ref 0
         let badCats, failStreams, timedOut = CatStats(), HashSet(), ref 0
         let toStreams, oStreams = HashSet(), HashSet()
@@ -112,7 +112,7 @@ module Internal =
 
     type EventStoreSchedulingEngine =
         static member Create(log : ILogger, storeLog, connections : _ [], itemDispatcher, stats : EventStoreStats, dumpStreams, ?maxBatches)
-            : Scheduling.StreamSchedulingEngine<_, _> =
+            : Scheduling.StreamSchedulingEngine<_, _, _> =
             let writerResultLog = log.ForContext<Writer.Result>()
             let mutable robin = 0
 
@@ -134,9 +134,9 @@ module Internal =
                     | Choice2Of2 (_stats, _exn) -> streams.SetMalformed(stream, false)
                 let _stream, ss = applyResultToStreamState res
                 Writer.logTo writerResultLog (stream, res)
-                ss.write
+                ss.write, res
 
-            let dispatcher = Scheduling.MultiDispatcher<_, _>(itemDispatcher, attemptWrite, interpretWriteResultProgress, stats, dumpStreams)
+            let dispatcher = Scheduling.MultiDispatcher<_, _, _>(itemDispatcher, attemptWrite, interpretWriteResultProgress, stats, dumpStreams)
             Scheduling.StreamSchedulingEngine(dispatcher, enableSlipstreaming = true, ?maxBatches = maxBatches, idleDelay = TimeSpan.FromMilliseconds 2.)
 
 type EventStoreSink =