+++ src/Propulsion.Kafka/Consumers.fs
@@ -206,11 +206,11 @@ type ParallelConsumer private () =
         ParallelConsumer.Start<KeyValuePair<string, string>>(log, config, maxDop, Bindings.mapConsumeResult, handle >> Async.Catch,
             ?maxSubmissionsPerPartition=maxSubmissionsPerPartition, ?pumpInterval=pumpInterval, ?statsInterval=statsInterval, ?logExternalStats=logExternalStats)
 
-type EventStats = Streams.EventStats
+type EventMetrics = Streams.EventMetrics
 
 [<AbstractClass>]
 type StreamsConsumerStats<'Outcome>(log : ILogger, statsInterval, stateInterval) =
-    inherit Streams.Scheduling.StreamSchedulerStats<EventStats * 'Outcome, EventStats * exn>(log, statsInterval, stateInterval)
+    inherit Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>(log, statsInterval, stateInterval)
     let okStreams, failStreams = HashSet(), HashSet()
     let mutable okEvents, okBytes, exnEvents, exnBytes = 0, 0L, 0, 0L
 
@@ -245,14 +245,14 @@ module Core =
 
         static member Start<'Info, 'Outcome>
             (   log : ILogger, config : KafkaConsumerConfig, resultToInfo, infoToStreamEvents,
-                prepare, handle, maxDop, stats : Streams.Scheduling.StreamSchedulerStats<int64 * EventStats * 'Outcome, EventStats * exn>,
+                prepare, handle, maxDop, stats : Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>,
                 ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay, ?maxBatches, ?maximizeOffsetWriting) =
             let pipelineStatsInterval = defaultArg pipelineStatsInterval (TimeSpan.FromMinutes 10.)
             let dispatcher = Streams.Scheduling.ItemDispatcher<_> maxDop
             let dumpStreams (streams : Streams.Scheduling.StreamStates<_>) log =
                 logExternalState |> Option.iter (fun f -> f log)
                 streams.Dump(log, Streams.Buffering.StreamState.eventsSize)
-            let streamsScheduler = Streams.Scheduling.StreamSchedulingEngine.Create<_, _, _>(dispatcher, stats, prepare, handle, dumpStreams, ?idleDelay=idleDelay, ?maxBatches=maxBatches)
+            let streamsScheduler = Streams.Scheduling.StreamSchedulingEngine.Create<_, _, _, _>(dispatcher, stats, prepare, handle, Streams.SpanResult.toIndex, dumpStreams, ?idleDelay=idleDelay, ?maxBatches=maxBatches)
             let mapConsumedMessagesToStreamsBatch onCompletion (x : Submission.SubmissionBatch<'Info>) : Streams.Scheduling.StreamsBatch<_> =
                 let onCompletion () = x.onCompletion(); onCompletion()
                 Streams.Scheduling.StreamsBatch.Create(onCompletion, Seq.collect infoToStreamEvents x.messages) |> fst
@@ -266,8 +266,8 @@ module Core =
 
         static member Start<'Info, 'Outcome>
             (   log : ILogger, config : KafkaConsumerConfig, consumeResultToInfo, infoToStreamEvents,
-                handle : StreamName * Streams.StreamSpan<_> -> Async<int64 * 'Outcome>, maxDop,
-                stats : Streams.Scheduling.StreamSchedulerStats<int64 * EventStats * 'Outcome, EventStats * exn>,
+                handle : StreamName * Streams.StreamSpan<_> -> Async<Streams.SpanResult * 'Outcome>, maxDop,
+                stats : Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>,
                 ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay, ?maxBatches, ?maximizeOffsetWriting) =
             let prepare (streamName, span) =
                 let stats = Streams.Buffering.StreamSpan.stats span
@@ -292,8 +292,8 @@ module Core =
             (   log : ILogger, config : KafkaConsumerConfig,
                 /// often implemented via <c>StreamNameSequenceGenerator.KeyValueToStreamEvent</c>
                 keyValueToStreamEvents,
-                prepare, handle : StreamName * Streams.StreamSpan<_> -> Async<int64 * 'Outcome>,
-                maxDop, stats : Streams.Scheduling.StreamSchedulerStats<int64 * EventStats * 'Outcome, EventStats * exn>,
+                prepare, handle : StreamName * Streams.StreamSpan<_> -> Async<Streams.SpanResult * 'Outcome>,
+                maxDop, stats : Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>,
                 ?maximizeOffsetWriting, ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay)=
             StreamsConsumer.Start<KeyValuePair<string, string>, 'Outcome>(
                 log, config, Bindings.mapConsumeResult, keyValueToStreamEvents, prepare, handle, maxDop, stats,
@@ -309,8 +309,8 @@ module Core =
             (   log : ILogger, config : KafkaConsumerConfig,
                 /// often implemented via <c>StreamNameSequenceGenerator.KeyValueToStreamEvent</c>
                 keyValueToStreamEvents : KeyValuePair<string, string> -> Propulsion.Streams.StreamEvent<_> seq,
-                handle : StreamName * Streams.StreamSpan<_> -> Async<int64 * 'Outcome>, maxDop,
-                stats : Streams.Scheduling.StreamSchedulerStats<int64 * EventStats * 'Outcome, EventStats * exn>,
+                handle : StreamName * Streams.StreamSpan<_> -> Async<Streams.SpanResult * 'Outcome>, maxDop,
+                stats : Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>,
                 ?maximizeOffsetWriting, ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay, ?maxBatches) =
             StreamsConsumer.Start<KeyValuePair<string, string>, 'Outcome>(
                 log, config, Bindings.mapConsumeResult, keyValueToStreamEvents, handle, maxDop, stats,
@@ -410,13 +410,13 @@ type StreamsConsumer =
             /// - second component: Outcome (can be simply <c>unit</c>), to pass to the <c>stats</c> processor
             /// - throwing marks the processing of a stream as having faulted (the stream's pending events and/or
             ///   new ones that arrived while the handler was processing are then eligible for retry purposes in the next dispatch cycle)
-            handle : StreamName * Streams.StreamSpan<_> -> Async<int64 * 'Outcome>,
+            handle : StreamName * Streams.StreamSpan<_> -> Async<Streams.SpanResult * 'Outcome>,
             /// The maximum number of instances of <c>handle</c> that are permitted to be dispatched at any point in time.
             /// The scheduler seeks to maximise the in-flight <c>handle</c>rs at any point in time.
             /// The scheduler guarantees to never schedule two concurrent <c>handler<c> invocations for the same stream.
             maxDop,
             /// The <c>'Outcome</c> from each handler invocation is passed to the Statistics processor by the scheduler for periodic emission
-            stats : Streams.Scheduling.StreamSchedulerStats<int64 *EventStats * 'Outcome, EventStats * exn>,
+            stats : Streams.Scheduling.Stats<EventMetrics * 'Outcome, EventMetrics * exn>,
             /// Prevent batches being consolidated prior to scheduling in order to maximize granularity of consumer offset updates
             ?maximizeOffsetWriting,
             ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay, ?maxBatches) =
@@ -451,7 +451,7 @@ type BatchesConsumer =
             ///   new ones that arrived while the handler was processing are then eligible for retry purposes in the next dispatch cycle)
             handle : Streams.Scheduling.DispatchItem<_>[] -> Async<seq<Choice<int64, exn>>>,
             /// The responses from each <c>handle</c> invocation are passed to <c>stats</c> for periodic emission
-            stats : Streams.Scheduling.StreamSchedulerStats<int64 * EventStats * unit, EventStats * exn>,
+            stats : Streams.Scheduling.Stats<EventMetrics * unit, EventMetrics * exn>,
             /// Maximum number of batches to ingest for scheduling at any one time (Default: 24.)
             /// NOTE Stream-wise consumption defaults to taking 5 batches each time replenishment is required
             ?schedulerIngestionBatchCount, ?pipelineStatsInterval, ?maxSubmissionsPerPartition, ?pumpInterval, ?logExternalState, ?idleDelay) =
@@ -461,7 +461,7 @@ type BatchesConsumer =
             logExternalState |> Option.iter (fun f -> f log)
             streams.Dump(log, Streams.Buffering.StreamState.eventsSize)
         let handle (items : Streams.Scheduling.DispatchItem<byte[]>[])
-            : Async<(StreamName * Choice<int64 * EventStats * unit, EventStats * exn>)[]> = async {
+            : Async<(StreamName * Choice<int64 * (EventMetrics * unit), EventMetrics * exn>)[]> = async {
             try let! results = handle items
                 return
                     [| for x in Seq.zip items results ->
@@ -469,7 +469,7 @@ type BatchesConsumer =
                         | item, Choice1Of2 index' ->
                             let used : Streams.StreamSpan<_> = { item.span with events = item.span.events |> Seq.takeWhile (fun e -> e.Index <> index' ) |> Array.ofSeq }
                             let s = Streams.Buffering.StreamSpan.stats used
-                            item.stream, Choice1Of2 (index', s, ())
+                            item.stream, Choice1Of2 (index', (s, ()))
                         | item, Choice2Of2 exn ->
                             let s = Streams.Buffering.StreamSpan.stats item.span
                             item.stream, Choice2Of2 (s, exn) |]